refactor(memory): move conversation memory config to routers/common
Move ConversationMemoryConfig, LongTermMemoryConfig, ShortTermMemoryConfig,
and extract_conversation_memory_config from responses/header_utils.rs to
routers/common/header_utils.rs where shared header parsing utilities live.
