+++ model_gateway/src/routers/openai/responses/mod.rs
@@ -11,7 +11,6 @@
 
 mod accumulator;
 mod common;
-pub(crate) mod header_utils;
 pub(crate) mod history;
 mod non_streaming;
 pub(crate) mod route;
